+++ concrete_datastore/api/v1/filters.py
@@ -83,9 +83,9 @@ def convert_type(string, field_type, close_period=True):
             # Expected format YYYY-MM-DD HH:MM[:ss[.uuuuuu]][TZ]
             dt = ensure_pendulum(string)
             if close_period:
-                dt.end_of('day')
+                dt = dt.end_of('day')
             else:
-                dt.start_of('day')
+                dt = dt.start_of('day')
             return format_datetime(dt)
         if field_type == 'DateField':
             # Expected YYYY-MM-DD
@@ -315,6 +315,51 @@ class FilterSupportingContainsBackend(
             if not param.endswith('__contains'):
                 continue
             param_field = param.replace('__contains', '')
+            if param_field not in filterset_fields:
+                continue
+            if not get_filter_field_type(queryset.model, param_field) in (
+                'CharField',
+                'TextField',
+            ):
+                continue
+
+            custom_filter = {param: query_params.get(param)}
+            if q_object is None:
+                q_object = Q(**custom_filter)
+            else:
+                q_object &= Q(**custom_filter)
+        if q_object is None:
+            return queryset
+
+        return queryset.filter(q_object)
+
+
+class FilterSupportingInsensitiveContainsBackend(
+    BaseFilterBackend, CustomShemaOperationParameters
+):
+    def get_schema_operation_parameters(self, view):
+        params = [
+            {
+                'name': f'{field_name}__icontains',
+                'required': False,
+                'in': 'query',
+                'schema': {'type': 'string'},
+            }
+            for field_name in getattr(view, 'filterset_fields', ())
+            if get_filter_field_type(view.model_class, field_name)
+            in ('CharField', 'TextField')
+        ]
+        return self.return_if_not_details(view=view, value=params)
+
+    def filter_queryset(self, request, queryset, view):
+        query_params = request.query_params
+        filterset_fields = getattr(view, 'filterset_fields', ())
+
+        q_object = None
+        for param in query_params:
+            if not param.endswith('__icontains'):
+                continue
+            param_field = param.replace('__icontains', '')
             if param_field not in filterset_fields:
                 continue
             if not get_filter_field_type(queryset.model, param_field) in (